+++ src/main/java/org/scion/multiping/PingRepeat.java
@@ -27,6 +27,7 @@ import java.util.*;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import org.scion.jpan.*;
 import org.scion.jpan.internal.PathRawParser;
+import org.scion.jpan.internal.Shim;
 import org.scion.multiping.util.*;
 import org.scion.multiping.util.Record;
 
@@ -54,6 +55,7 @@ public class PingRepeat {
   private int nPingSuccess = 0;
   private int nPingTimeout = 0;
   private int nPingError = 0;
+  private static int localPort = -1;
 
   private static Config config;
   private static FileWriter fileWriter;
@@ -65,10 +67,13 @@ public class PingRepeat {
   }
 
   public static void main(String[] args) throws IOException {
-    // System.setProperty(Constants.PROPERTY_DNS_SEARCH_DOMAINS, "ethz.ch.");
-
     config = Config.read(FILE_CONFIG);
     PRINT = config.consoleOutput;
+
+    localPort = config.hasLocalPort() ? config.localPort : -1;
+    println("Settings");
+    println(" Listening on port: " + localPort);
+    println(" JPAN SHIM is running: " + Shim.isInstalled());
 
     // Output: ISD/AS, remote IP, time, hopCount, path, [pings]
     fileWriter = new FileWriter(config.outputFile);
@@ -141,7 +146,12 @@ public class PingRepeat {
     // output
     int nHops = PathRawParser.create(rec.getPath().getRawPath()).getHopCount();
     String out = rec.getRemoteIP() + "  nPaths=" + nPaths + "  nHops=" + nHops;
-    out += "  time=" + bestAttempt.get().getPingMs() + "ms" + "  ICMP=" + icmpMs;
+    if (bestAttempt.get().getState() == Record.Attempt.State.SUCCESS) {
+      out += "  time=" + bestAttempt.get().getPingMs() + "ms";
+    } else {
+      out += "  time=" + bestAttempt.get().getState();
+    }
+    out += "  ICMP=" + icmpMs;
     if (SHOW_PATH) {
       out += "  " + ScionUtil.toStringPath(rec.getPath().getMetadata());
     }
@@ -161,7 +171,8 @@ public class PingRepeat {
     Record best = null;
     double currentBestMs = Double.MAX_VALUE;
     ResponseHandler handler = new ResponseHandler();
-    try (ScmpSenderAsync sender = Scmp.newSenderAsyncBuilder(handler).build()) {
+    try (ScmpSenderAsync sender =
+        Scmp.newSenderAsyncBuilder(handler).setLocalPort(localPort).build()) {
       for (int attemptCount = 0; attemptCount < config.attemptRepeatCnt; attemptCount++) {
         Instant start = Instant.now();
         Map<Integer, Record> seqToPathMap = new HashMap<>();